telegramgptbot. temperature как double

+++ src/main/java/nikonov/telegramgptbot/external/gpt/dto/GPTRequest.java
@@ -22,5 +22,5 @@ public class GPTRequest {
     private int maxTokens;
     
     @JsonProperty("temperature")
-    private int temperature;
+    private double temperature;
 }
